fix: move 7 test files to correct sibling locations
These tests were placed at the feature root instead of next to their
actual source file in a subdirectory:

- SliderInput.test.tsx → components/controls/
- SnappingSlider.test.tsx → components/controls/
- generateSuggestions.test.ts → utils/
- fileNaming.test.ts → utils/
- layerAutoExpansion.test.ts → utils/
- layoutBuilder.test.ts → utils/
- storage.test.ts → storage/DesignerStorage.test.ts
