+++ src/features/name-suggestions/utils/generateSuggestions.test.ts
@@ -3,9 +3,9 @@
  */
 
 import { describe, it, expect } from 'vitest';
-import { generateSuggestions } from './utils/generateSuggestions';
-import { hashName, editDistance } from './utils/stringUtils';
-import type { SuggestionInput } from './types';
+import { generateSuggestions } from './generateSuggestions';
+import { hashName, editDistance } from './stringUtils';
+import type { SuggestionInput } from '../types';
 
 describe('generateSuggestions', () => {
   const baseInput: SuggestionInput = {
